Disassembler: Get rid of PAGE_SIZE dependency in userspace for hosts where the page size can't be determined during build time. Use the smallest possible page size available across all supported architectures, bugref:10391 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108700 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Disassembler/Disasm.cpp
@@ -55,7 +55,7 @@ static DECLCALLBACK(int) disReadBytesDefault(PDISSTATE pDis, uint8_t offInstr, u
 #ifdef IN_RING0
     size_t  const   cbLeftOnPage = (uintptr_t)pbSrc & PAGE_OFFSET_MASK;
 #else
-    size_t  const   cbLeftOnPage = (uintptr_t)pbSrc & (_4K - 1); /* Minimum page size for safety. */
+    size_t  const   cbLeftOnPage = (uintptr_t)pbSrc & (uintptr_t)(_4K - 1); /* Minimum page size for safety. */
 #endif
     uint8_t         cbToRead     = cbLeftOnPage >= cbMaxRead
                                  ? cbMaxRead
